VDB-5230 added test

+++ test/external/vdb-validate/vdb_validate.sh
@@ -79,5 +79,12 @@ mkdir -p actual/
 	 then echo ${vdb_validate} no-schema-run should fail; exit 1; fi
 	grep -q ' Run File is obsolete. Please download the latest' actual/noschema
 
+	# no error message (PLATFORM column undefined) issued if run on a reference sequence
+	output=$(./runtestcase.sh "${bin_dir}/${vdb_validate} db/no-platform" no-platform 0)
+	res=$?
+	if [ "$res" != "0" ];
+		then echo "${vdb_validate} FAILED, res=$res output=$output" && exit 1;
+	fi
+
 	echo "All ${vdb_validate} tests succeed"
 	rm -rf actual/
